move event related code into separate files
