+++ src/cards/side.ts
@@ -1,0 +1,6 @@
+enum Side {
+  Front = "FRONT",
+  Back = "BACK",
+}
+
+export default Side;
