+++ src/index.tsx
@@ -3,6 +3,8 @@ import * as ReactDOM from "react-dom";
 import { range, rangeFromTo } from "lang/ranges";
 import { removeFromArray, shuffle } from "lang/arrays";
 import { noop, NoProps, NoState } from "lang/react";
+import StackPopped from "stacks/stackPopped";
+import BringEventToConclusion from "events/bringEventToConclusion";
 import PartyBattleGoals from "battlegoals/partyBattleGoals";
 
 class RandomSideScenarioProps {
@@ -216,142 +218,6 @@ class AddCards extends React.Component<AddCardsProps, NoState> {
         <input ref={(i) => this.inputs[cardType] = i}/>
       </div>
     ));
-  }
-}
-
-
-enum Side {
-  Front = "FRONT",
-  Back = "BACK",
-}
-
-interface EventCardProps {
-  eventCardId: number,
-  name: string,
-  side: Side
-}
-
-class EventCard extends React.Component<EventCardProps, NoState> {
-  constructor(props: EventCardProps) {
-    super(props);
-    this.eventCardImageUrl = this.eventCardImageUrl.bind(this);
-  }
-
-  render() {
-    return <img key='image-front' src={this.eventCardImageUrl()}/>
-  }
-
-  eventCardImageUrl() {
-    const twoDigitNumber = (this.props.eventCardId <= 9 ? "0" : "") + this.props.eventCardId;
-    const imageName = this.props.name.toLowerCase();
-    const imageBaseUrl = "https://raw.githubusercontent.com/any2cards/gloomhaven/master/images/events/base/" + imageName + "/" + imageName.charAt(0) + "e-" + twoDigitNumber + "-";
-    const sideUrlPart = this.props.side === Side.Back ? 'b' : 'f';
-    return imageBaseUrl + sideUrlPart + '.png';
-  }
-}
-
-interface StackPopped {
-  stackPopped: (name: string, returnToBottom: boolean) => void,
-}
-
-interface ButtonWithSelectionHighlightProps {
-  onClick: React.MouseEventHandler;
-  selected: boolean;
-  text: string;
-}
-
-class ButtonWithSelectionHighlight extends React.Component<ButtonWithSelectionHighlightProps, NoState> {
-  render() {
-    const style = {} as React.CSSProperties;
-    if (this.props.selected) {
-      style.borderColor = 'red';
-    }
-    return <button key={this.props.text} style={style} type="button" onClick={this.props.onClick}>{this.props.text}</button>;
-  }
-}
-
-interface BringEventToConclusionProps extends StackPopped{
-  name: string,
-  number: number,
-}
-
-enum Choice {
-  A = "A",
-  B = "B",
-}
-
-interface BringEventToConclusionState {
-  selected: Choice | undefined
-}
-
-class BringEventToConclusion extends React.Component<BringEventToConclusionProps, BringEventToConclusionState> {
-  constructor(props: BringEventToConclusionProps) {
-    super(props);
-
-    this.selectA = this.selectA.bind(this);
-    this.selectB = this.selectB.bind(this);
-    this.returnToBottom = this.returnToBottom.bind(this);
-    this.removeFromGame = this.removeFromGame.bind(this);
-    this.state = { selected: undefined };
-  }
-
-  selectA(e: React.MouseEvent) {
-    this.setState({selected: Choice.A});
-    e.preventDefault();
-    return false;
-  }
-
-  selectB(e: React.MouseEvent) {
-    this.setState({selected: Choice.B});
-    e.preventDefault();
-    return false;
-  }
-
-  returnToBottom(e: React.MouseEvent) {
-    this.props.stackPopped(this.props.name, true);
-    e.preventDefault();
-    return false;
-  }
-
-  removeFromGame(e: React.MouseEvent) {
-    this.props.stackPopped(this.props.name, false);
-    e.preventDefault();
-    return false;
-  }
-
-  render() {
-    const containerStyle = {
-      'display': 'flex',
-      'flex-direction': 'row'
-    };
-    const choiceStyle = {
-      'display': 'flex',
-      'flex-direction': 'column',
-      'padding': '0 1em 0'
-    };
-    const resolutionStyle = {
-      'display': 'flex',
-      'flex-direction': 'column',
-      'padding': '0 1em 0 '
-    };
-    return [<h2 key="h2">{this.props.name} Event {this.props.number}</h2>,
-      <div style={containerStyle}>
-        <EventCard key='event-card-front' eventCardId={this.props.number} side={Side.Front} name={this.props.name}/>
-        <div style={choiceStyle}>
-          <ButtonWithSelectionHighlight onClick={this.selectA} selected={this.state.selected === Choice.A} text="A"/>
-          <ButtonWithSelectionHighlight onClick={this.selectB} selected={this.state.selected === Choice.B} text="B"/>
-        </div>
-        {this.state.selected && [
-          <EventCard key='event-card-back' eventCardId={this.props.number} side={Side.Back} name={this.props.name}/>,
-          <div style={resolutionStyle}>
-            {this.props.children}
-            <h3>Conclusion</h3>
-            <button key="ret" type="button" onClick={this.returnToBottom}>Return to bottom</button>
-            <button key="rem" type="button" onClick={this.removeFromGame}>Remove from game</button>
-          </div>
-        ]}
-      </div>
-    ];
   }
 }
