+++ src/stacks/stackPopped.ts
@@ -1,0 +1,4 @@
+
+export default interface StackPopped {
+  stackPopped: (name: string, returnToBottom: boolean) => void,
+}
